# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CoffeeCom Ecommerce API",
    "version": "1.0.0",
    "description": "Full OpenAPI 3.0 specification for CoffeeCom Flask backend. Import into Apidog or Postman to test all endpoints."
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT token" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "Get user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile info" },
          "401": { "description": "Unauthorized" }
        }
      },
      "put": {
        "summary": "Update user profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "username": { "type": "string" },
                  "address": { "type": "string" },
                  "phone_number": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated" },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "summary": "Delete user account",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Account deleted" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}